+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added support for `halfvec` and `sparsevec` types to Psycopg 3
 - Added `L1Distance` for Django
-- Added `l1_distance` for SQLAlchemy and SQLModel
+- Added `l1_distance` for SQLAlchemy, SQLModel, and Peewee
 
 ## 0.2.5 (2024-02-07)